+++ tests/integration_tests/test_model_exchange_coinbase_pro.py
@@ -3,8 +3,8 @@ from datetime import datetime
 
 sys.path.append('')
 # pylint: disable=import-error
-from models.exchange.coinbase_pro import AuthAPI, PublicAPI
-from models.helper.LogHelper import Logger
+from pycryptobot.models.exchange.coinbase_pro import AuthAPI, PublicAPI
+from pycryptobot.models.helper.LogHelper import Logger
 Logger.configure()
 
 DEFAULT_ORDER_MARKET = 'BTC-GBP'
