+++ tests/unit_tests/test_binance.py.skip
@@ -14,7 +14,7 @@ urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 
 sys.path.append('.')
 # pylint: disable=import-error
-from models.exchange.binance import AuthAPI, PublicAPI
+from pycryptobot.models.exchange.binance import AuthAPI, PublicAPI
 
 
 def test_instantiate_authapi_without_error():
@@ -112,7 +112,7 @@ def test_get_account(mocker):
     exchange = AuthAPI(api_key, api_secret, api_url)
     assert type(exchange) is AuthAPI
 
-    mocker.patch("models.exchange.binance.getAccount", return_value=client_response)
+    mocker.patch("pycryptobot.models.exchange.binance.getAccount", return_value=client_response)
     df = exchange.getAccount()
     assert type(df) is pandas.core.frame.DataFrame
 
@@ -144,7 +144,7 @@ def test_get_fees_with_market(mocker):
     exchange = AuthAPI(api_key, api_secret, api_url)
     assert type(exchange) is AuthAPI
 
-    mocker.patch("models.exchange.binance.Client.get_trade_fee", return_value=client_response)
+    mocker.patch("pycryptobot.models.exchange.binance.Client.get_trade_fee", return_value=client_response)
     df = exchange.getFees(MOCK_MARKET)
     assert type(df) is pandas.core.frame.DataFrame
 
@@ -178,7 +178,7 @@ def test_get_taker_fee_with_market(mocker):
     exchange = AuthAPI(api_key, api_secret, api_url)
     assert type(exchange) is AuthAPI
 
-    mocker.patch("models.exchange.binance.Client.get_trade_fee", return_value=client_response)
+    mocker.patch("pycryptobot.models.exchange.binance.Client.get_trade_fee", return_value=client_response)
     fee = exchange.getTakerFee(MOCK_MARKET)
     assert type(fee) is float
     assert fee == 0.001
@@ -206,7 +206,7 @@ def test_get_maker_fee_with_market(mocker):
     exchange = AuthAPI(api_key, api_secret, api_url)
     assert type(exchange) is AuthAPI
 
-    mocker.patch("models.exchange.binance.Client.get_trade_fee", return_value=client_response)
+    mocker.patch("pycryptobot.models.exchange.binance.Client.get_trade_fee", return_value=client_response)
     fee = exchange.getMakerFee(MOCK_MARKET)
     assert type(fee) is float
     assert fee == 0.001
@@ -256,7 +256,7 @@ def test_get_orders(mocker):
 
     exchange = AuthAPI(api_key, api_secret, api_url)
     assert type(exchange) is AuthAPI
-    mocker.patch("models.exchange.binance.Client.get_all_orders", return_value=client_response)
+    mocker.patch("pycryptobot.models.exchange.binance.Client.get_all_orders", return_value=client_response)
     df = exchange.getOrders(MOCK_MARKET)
 
     assert len(df) > 0
